+++ pkg/relay/relay.go
@@ -1,11 +1,11 @@
-//go:generate mockgen  -destination=./mocks/mocks.go -package=mocks github.com/blocknative/dreamboat/pkg/relay Datastore,State
+//go:generate mockgen  -destination=./mocks/mocks.go -package=mocks github.com/blocknative/dreamboat/pkg/relay Datastore,State,ValidatorStore,ValidatorCache
 package relay
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
+	"sync"
 	"time"
 
 	"github.com/flashbots/go-boost-utils/bls"
@@ -35,6 +35,15 @@ var (
 	ErrInvalidSlot             = errors.New("invalid slot")
 	ErrEmptyBlock              = errors.New("block is empty")
 )
+
+type ValidatorStore interface {
+	GetRegistration(context.Context, structs.PubKey) (types.SignedValidatorRegistration, error)
+}
+
+type ValidatorCache interface {
+	Add(types.PublicKey, structs.ValidatorCacheEntry) (evicted bool)
+	Get(types.PublicKey) (structs.ValidatorCacheEntry, bool)
+}
 
 type State interface {
 	Beacon() *structs.BeaconState
@@ -93,22 +102,31 @@ type Relay struct {
 	ver    Verifier
 	config RelayConfig
 
+	cache  ValidatorCache
+	vstore ValidatorStore
+
 	beacon      Beacon
 	beaconState State
+
+	deliveredCache     map[uint64]struct{}
+	deliveredCacheLock sync.RWMutex
 
 	m RelayMetrics
 }
 
 // NewRelay relay service
-func NewRelay(l log.Logger, config RelayConfig, beacon Beacon, ver Verifier, beaconState State, d Datastore, a Auctioneer) *Relay {
+func NewRelay(l log.Logger, config RelayConfig, beacon Beacon, cache ValidatorCache, vstore ValidatorStore, ver Verifier, beaconState State, d Datastore, a Auctioneer) *Relay {
 	rs := &Relay{
-		d:           d,
-		a:           a,
-		l:           l,
-		ver:         ver,
-		config:      config,
-		beacon:      beacon,
-		beaconState: beaconState,
+		d:              d,
+		a:              a,
+		l:              l,
+		ver:            ver,
+		config:         config,
+		cache:          cache,
+		vstore:         vstore,
+		beacon:         beacon,
+		beaconState:    beaconState,
+		deliveredCache: make(map[uint64]struct{}),
 	}
 	rs.initMetrics()
 	return rs
@@ -304,213 +322,4 @@ func (rs *Relay) GetPayload(ctx context.Context, m *structs.MetricGroup, payload
 		Version: "bellatrix",
 		Data:    payload.Payload.Data,
 	}, nil
-}
-
-// ***** Relay Domain *****
-// SubmitBlockRequestToSignedBuilderBid converts a builders block submission to a bid compatible with mev-boost
-func SubmitBlockRequestToSignedBuilderBid(req *types.BuilderSubmitBlockRequest, sk *bls.SecretKey, pubkey *types.PublicKey, domain types.Domain) (*types.SignedBuilderBid, error) { // TODO(l): remove FB type
-	if req == nil {
-		return nil, ErrMissingRequest
-	}
-
-	if sk == nil {
-		return nil, ErrMissingSecretKey
-	}
-
-	header, err := types.PayloadToPayloadHeader(req.ExecutionPayload)
-	if err != nil {
-		return nil, err
-	}
-
-	builderBid := types.BuilderBid{
-		Value:  req.Message.Value,
-		Header: header,
-		Pubkey: *pubkey,
-	}
-
-	sig, err := types.SignMessage(&builderBid, domain, sk)
-	if err != nil {
-		return nil, err
-	}
-
-	return &types.SignedBuilderBid{
-		Message:   &builderBid,
-		Signature: sig,
-	}, nil
-}
-
-// SubmitBlock Accepts block from trusted builder and stores
-func (rs *Relay) SubmitBlock(ctx context.Context, m *structs.MetricGroup, submitBlockRequest *types.BuilderSubmitBlockRequest) error {
-	tStart := time.Now()
-	defer m.AppendSince(tStart, "submitBlock", "all")
-
-	logger := rs.l.With(log.F{
-		"method":    "SubmitBlock",
-		"builder":   submitBlockRequest.Message.BuilderPubkey,
-		"blockHash": submitBlockRequest.ExecutionPayload.BlockHash,
-		"slot":      submitBlockRequest.Message.Slot,
-		"proposer":  submitBlockRequest.Message.ProposerPubkey,
-		"bid":       submitBlockRequest.Message.Value.String(),
-	})
-
-	_, err := rs.verifyBlock(submitBlockRequest, rs.beaconState.Beacon())
-	if err != nil {
-		return fmt.Errorf("%w: %s", ErrVerification, err.Error()) // TODO: multiple err wrapping in Go 1.20
-	}
-
-	tCheckDelivered := time.Now()
-	slot := structs.Slot(submitBlockRequest.Message.Slot)
-	ok, err := rs.d.CheckSlotDelivered(ctx, uint64(slot))
-	m.AppendSince(tCheckDelivered, "submitBlock", "checkDelivered")
-	if ok {
-		return ErrPayloadAlreadyDelivered
-	}
-	if err != nil {
-		return err
-	}
-
-	tVerify := time.Now()
-	msg, err := types.ComputeSigningRoot(submitBlockRequest.Message, rs.config.BuilderSigningDomain)
-	if err != nil {
-		return ErrInvalidSignature
-	}
-
-	err = rs.ver.Enqueue(ctx, submitBlockRequest.Signature, submitBlockRequest.Message.BuilderPubkey, msg)
-	m.AppendSince(tVerify, "submitBlock", "verify")
-
-	if err != nil {
-		return fmt.Errorf("%w: %s", ErrVerification, err.Error()) // TODO: multiple err wrapping in Go 1.20
-	}
-
-	complete, err := rs.prepareContents(submitBlockRequest)
-	if err != nil {
-		return fmt.Errorf("fail to generate contents from block submission: %w", err)
-	}
-
-	b, err := json.Marshal(complete.Header)
-	if err != nil {
-		return fmt.Errorf("%w block as header: %s", ErrMarshal, err.Error()) // TODO: multiple err wrapping in Go 1.20
-	}
-
-	tPutPayload := time.Now()
-	if err := rs.d.PutPayload(ctx, SubmissionToKey(submitBlockRequest), &complete.Payload, rs.config.TTL); err != nil {
-		return fmt.Errorf("%w block as payload: %s", ErrStore, err.Error()) // TODO: multiple err wrapping in Go 1.20
-	}
-	m.AppendSince(tPutPayload, "submitBlock", "putPayload")
-
-	tAddAuction := time.Now()
-	isNewMax := rs.a.AddBlock(&complete)
-	m.AppendSince(tAddAuction, "submitBlock", "addAuction")
-
-	tPutHeader := time.Now()
-	err = rs.d.PutHeader(ctx, structs.HeaderData{
-		Slot:           slot,
-		Marshaled:      b,
-		HeaderAndTrace: complete.Header,
-	}, rs.config.TTL)
-	if err != nil {
-		return fmt.Errorf("%w block as header: %s", ErrStore, err.Error()) // TODO: multiple err wrapping in Go 1.20
-	}
-	m.AppendSince(tPutHeader, "submitBlock", "putHeader")
-
-	logger.With(log.F{
-		"processingTimeMs": time.Since(tStart).Milliseconds(),
-		"is_new_max":       isNewMax,
-	}).Trace("builder block stored")
-
-	return nil
-}
-
-func (rs *Relay) prepareContents(submitBlockRequest *types.BuilderSubmitBlockRequest) (structs.CompleteBlockstruct, error) {
-	s := structs.CompleteBlockstruct{}
-
-	signedBuilderBid, err := SubmitBlockRequestToSignedBuilderBid(
-		submitBlockRequest,
-		rs.config.SecretKey,
-		&rs.config.PubKey,
-		rs.config.BuilderSigningDomain,
-	)
-	if err != nil {
-		return s, err
-	}
-
-	s.Payload = SubmitBlockRequestToBlockBidAndTrace(signedBuilderBid, submitBlockRequest)
-
-	header, err := types.PayloadToPayloadHeader(submitBlockRequest.ExecutionPayload)
-	if err != nil {
-		return s, err
-	}
-
-	s.Header = structs.HeaderAndTrace{
-		Header: header,
-		Trace: &structs.BidTraceWithTimestamp{
-			BidTraceExtended: structs.BidTraceExtended{
-				BidTrace: types.BidTrace{
-					Slot:                 submitBlockRequest.Message.Slot,
-					ParentHash:           s.Payload.Payload.Data.ParentHash,
-					BlockHash:            s.Payload.Payload.Data.BlockHash,
-					BuilderPubkey:        s.Payload.Trace.Message.BuilderPubkey,
-					ProposerPubkey:       s.Payload.Trace.Message.ProposerPubkey,
-					ProposerFeeRecipient: s.Payload.Trace.Message.ProposerFeeRecipient,
-					Value:                submitBlockRequest.Message.Value,
-					GasLimit:             s.Payload.Trace.Message.GasLimit,
-					GasUsed:              s.Payload.Trace.Message.GasUsed,
-				},
-				BlockNumber: s.Payload.Payload.Data.BlockNumber,
-				NumTx:       uint64(len(s.Payload.Payload.Data.Transactions)),
-			},
-			Timestamp:   uint64(time.Now().UnixMilli() / 1_000),
-			TimestampMs: uint64(time.Now().UnixMilli()),
-		},
-	}
-
-	return s, nil
-}
-
-func (rs *Relay) verifyBlock(submitBlockRequest *types.BuilderSubmitBlockRequest, beaconState *structs.BeaconState) (bool, error) { // TODO(l): remove FB type
-	if submitBlockRequest == nil || submitBlockRequest.Message == nil {
-		return false, ErrEmptyBlock
-	}
-
-	expectedTimestamp := beaconState.GenesisTime + (submitBlockRequest.Message.Slot * 12)
-	if submitBlockRequest.ExecutionPayload.Timestamp != expectedTimestamp {
-		return false, fmt.Errorf("%w: got %d, expected %d", ErrInvalidTimestamp, submitBlockRequest.ExecutionPayload.Timestamp, expectedTimestamp)
-	}
-
-	if structs.Slot(submitBlockRequest.Message.Slot) < beaconState.CurrentSlot {
-		return false, fmt.Errorf("%w: got %d, expected %d", ErrInvalidSlot, submitBlockRequest.Message.Slot, beaconState.CurrentSlot)
-	}
-
-	return true, nil
-}
-
-func SubmissionToKey(submission *types.BuilderSubmitBlockRequest) structs.PayloadKey {
-	return structs.PayloadKey{
-		BlockHash: submission.ExecutionPayload.BlockHash,
-		Proposer:  submission.Message.ProposerPubkey,
-		Slot:      structs.Slot(submission.Message.Slot),
-	}
-}
-
-func SubmitBlockRequestToBlockBidAndTrace(signedBuilderBid *types.SignedBuilderBid, submitBlockRequest *types.BuilderSubmitBlockRequest) structs.BlockBidAndTrace { // TODO(l): remove FB type
-	getHeaderResponse := types.GetHeaderResponse{
-		Version: "bellatrix",
-		Data:    signedBuilderBid,
-	}
-
-	getPayloadResponse := types.GetPayloadResponse{
-		Version: "bellatrix",
-		Data:    submitBlockRequest.ExecutionPayload,
-	}
-
-	signedBidTrace := types.SignedBidTrace{
-		Message:   submitBlockRequest.Message,
-		Signature: submitBlockRequest.Signature,
-	}
-
-	return structs.BlockBidAndTrace{
-		Trace:   &signedBidTrace,
-		Bid:     &getHeaderResponse,
-		Payload: &getPayloadResponse,
-	}
 }
